refactor: renamed popUpSelection to ExerciseSelectionPanel and split ExerciseSelectionPanel into modular components with context integration
