+++ src/components/ExerciseSelectionPanel/ExerciseSelectionPanel.css
